requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
apparmor==3.0.7
appdirs==1.4.4
application-utility==1.3.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.7
asttokens==2.0.7
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==2.6.0
black==22.6.0
bleach==5.0.1
btrfsutil==5.18.1
CacheControl==0.12.6
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contextlib2==21.6.0
cryptography==37.0.4
cupshelpers==1.0
cycler==0.11.0
cytoolz==0.12.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.5.1
distlib==0.3.5
distro==1.7.0
dnspython==2.2.1
docopt==0.6.2
docutils==0.19
entrypoints==0.4
eth-abi==2.2.0
eth-account==0.5.8
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
executing==0.10.0
fastjsonschema==2.16.1
fonttools==4.37.1
frozenlist==1.3.1
future==0.18.2
greenlet==1.1.2
gufw==22.4.0
hexbytes==0.2.2
html5lib==1.1
idna==3.3
imageio==2.21.2
importlib-metadata==4.8.1
ipfshttpclient==0.8.0a2
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isort==5.10.1
jaraco.context==4.1.1
jaraco.functools==3.5.1
jaraco.text==3.8.1
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.9
jsonpointer==2.3
jsonschema==4.9.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-packaging==0.12.2
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
keras==2.10.0
keyutils==0.6
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
lensfun==0.3.3
LibAppArmor==3.0.7
libtorrent==2.0.7
lit==14.0.6.dev0
louis==3.22.0
lru-dict==1.1.8
mahotas==1.4.13
manjaro-sdk==0.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
more-itertools==8.13.0
msgpack==1.0.3
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.4
netaddr==0.8.0
netsnmp-python==1.0a1
networkx==2.8.6
nftables==0.1
notebook==6.4.12
notebook-shim==0.1.0
npyscreen==4.10.5
nspektr==0.4.0
numpy==1.23.2
opencv-python==4.6.0.66
ordered-set==4.0.2
packaging==21.3
pacman-mirrors==4.23.2
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
ply==3.11
progress==1.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.21.0
pycparser==2.21
pycryptodome==3.15.0
pycups==2.0.1
pycurl==7.45.1
pygame==2.1.2
Pygments==2.12.0
PyGObject==3.42.2
pylint==2.14.5
pynvim==0.4.3
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.18.1
PySide6==6.3.1
pysmbc==1.0.23
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
rados==2.0.0
rbd==2.0.0
reportlab==3.6.11
requests==2.28.1
requests-unixsocket==0.3.0
resolvelib==0.8.1
retrying==1.3.3
rfc3987==1.3.8
rgw==2.0.0
rich @ file:///build/python-rich/src/rich-12.5.1/dist/rich-12.5.0-py3-none-any.whl
rlp==2.0.1
scikit-image==0.19.3
scipy==1.9.1
Send2Trash==1.8.0
shiboken6==6.3.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.4.0
systemd-python==234
TBB==0.2
team==1.0
tenacity==8.0.1
terminado==0.15.0
testpath==0.6.0
tifffile==2022.8.12
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
toolz==0.12.0
tornado==6.1
traitlets==5.3.0
trove-classifiers==2022.8.7
typing_extensions==4.3.0
udiskie==2.4.2
ufw==0.36.1
urllib3==1.26.11
validate-pyproject==0.9
varint==1.0.2
wcwidth==0.2.5
web3==5.30.0
webcolors==1.12
webencodings==0.5.1
websocket-client==1.3.3
websockets==9.1
widgetsnbextension==3.6.1
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
